src/requirements.txt
numpy==1.13.3
pdf2image==1.4.1
img2pdf==0.3.3
image_slicer==0.2.0
Pillow==5.4.1
PyPDF2==1.26.0
joblib==0.13.1
